[feat/mypage]: move GridSpacingItemDecoration to core/common
