+++ core/common/src/main/java/com/teampophory/pophory/common/view/ViewExt.kt
@@ -1,7 +1,9 @@
 package com.teampophory.pophory.common.view
 
+import android.graphics.Rect
 import android.view.View
 import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.RecyclerView
 
 inline fun View.setOnSingleClickListener(
     delay: Long = 500L,
@@ -28,4 +30,28 @@ class ItemDiffCallback<T: Any>(
     override fun areContentsTheSame(
         oldItem: T, newItem: T
     ): Boolean = onContentsTheSame(oldItem, newItem)
+}
+
+class GridSpacingItemDecoration(private val spanCount: Int, private val spacing: Int, private val includeEdge: Boolean) : RecyclerView.ItemDecoration() {
+
+    override fun getItemOffsets(outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State) {
+        val position = parent.getChildAdapterPosition(view) // item position
+        val column = position % spanCount // item column
+
+        if (includeEdge) {
+            outRect.left = spacing - column * spacing / spanCount // spacing - column * ((1f / spanCount) * spacing)
+            outRect.right = (column + 1) * spacing / spanCount // (column + 1) * ((1f / spanCount) * spacing)
+
+            if (position < spanCount) { // top edge
+                outRect.top = spacing
+            }
+            outRect.bottom = spacing // item bottom
+        } else {
+            outRect.left = column * spacing / spanCount // column * ((1f / spanCount) * spacing)
+            outRect.right = spacing - (column + 1) * spacing / spanCount // spacing - (column + 1) * ((1f / spanCount) * spacing)
+            if (position >= spanCount) {
+                outRect.top = spacing // item top
+            }
+        }
+    }
 }
